:truck: api: move jury certification serializer to src
