+++ api/src/certification/session-management/application/jury-certification-controller.js
@@ -1,5 +1,5 @@
-import * as juryCertificationSerializer from '../../../../lib/infrastructure/serializers/jsonapi/jury-certification-serializer.js';
 import { usecases } from '../domain/usecases/index.js';
+import * as juryCertificationSerializer from '../infrastructure/serializers/jury-certification-serializer.js';
 
 const getJuryCertification = async function (request, h, dependencies = { juryCertificationSerializer }) {
   const certificationCourseId = request.params.id;
