+++ api/src/certification/session-management/infrastructure/serializers/jury-certification-serializer.js
